pyproject.toml
[project]
name = "meowai-home"
version = "0.1.0"
description = "温馨的流浪猫AI收容所"
authors = [{name = "首席铲屎官"}]
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.19.0",
    "rich>=13.0.0",
    "nest_asyncio>=1.5.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "pydantic>=2.0.0",
    "pyjwt>=2.8.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
